Classifying trade on getTrades service and storying type on Trade
