+++ src/api/exchange/ExchangeApi.ts
@@ -148,6 +148,8 @@ export interface EventWithBlockInfo extends BaseTradeEvent {
   timestamp: number
 }
 
+export type TradeType = 'full' | 'partial' | 'liquidity' | 'unknown'
+
 /**
  * Trade enriches BaseTradeEvent with block, order and token data
  */
@@ -157,6 +159,7 @@ export interface Trade extends EventWithBlockInfo {
   settlingTimestamp: number
   buyToken: TokenDetails
   sellToken: TokenDetails
+  type?: TradeType
   limitPrice?: BigNumber
   fillPrice: BigNumber
   remainingAmount?: BN
