+++ src/utils/index.ts
@@ -7,3 +7,4 @@ export * from './price'
 export * from './validation'
 export * from './flagCodes'
 export * from './deferred'
+export * from './classifiers'
